fix: resolve Hibernate schema validation and test compilation errors
Map day_of_week as short to match SMALLINT column. Fix
InstitutionControllerTest for new DTO fields (laboratoryCount,
equipmentCount). Update LaboratoryServiceTest for removed
LaboratoryEquipmentRepository dependency and findAllWithCounts.

+++ server/src/main/java/com/example/carboncalculator/entities/LaboratorySchedule.java
@@ -42,7 +42,7 @@ public class LaboratorySchedule {
     private Laboratory laboratory;
 
     @Column(name = "day_of_week", nullable = false)
-    private int dayOfWeek;
+    private short dayOfWeek;
 
     @Column(name = "start_time", nullable = false)
     private LocalTime startTime;
